split eea and besu methods into 2 separate classes
