+++ acceptance-tests/src/test/java/tech/pegasys/ethsigner/tests/dsl/node/BesuNode.java
@@ -17,6 +17,7 @@ import static tech.pegasys.ethsigner.tests.WaitUtils.waitFor;
 
 import tech.pegasys.ethsigner.tests.dsl.Accounts;
 import tech.pegasys.ethsigner.tests.dsl.Besu;
+import tech.pegasys.ethsigner.tests.dsl.Eea;
 import tech.pegasys.ethsigner.tests.dsl.Eth;
 import tech.pegasys.ethsigner.tests.dsl.PrivateContracts;
 import tech.pegasys.ethsigner.tests.dsl.PublicContracts;
@@ -113,10 +114,11 @@ public class BesuNode implements Node {
     final RawJsonRpcRequestFactory requestFactory = new RawJsonRpcRequestFactory(web3jHttpService);
     final JsonRpc2_0Besu besuJsonRpc = new JsonRpc2_0Besu(web3jHttpService);
     final Eth eth = new Eth(jsonRpc);
-    final Besu besu = new Besu(besuJsonRpc, requestFactory);
+    final Besu besu = new Besu(besuJsonRpc);
+    final Eea eea = new Eea(requestFactory);
     this.accounts = new Accounts(eth);
     this.publicContracts = new PublicContracts(eth);
-    this.privateContracts = new PrivateContracts(besu);
+    this.privateContracts = new PrivateContracts(besu, eea);
     this.transactions = new Transactions(eth);
     this.ports = new NodePorts(httpRpcPort, wsRpcPort);
   }
