+++ ethsigner/core/src/test/java/tech/pegasys/ethsigner/core/jsonrpcproxy/EthAccountsBodyProviderTest.java
@@ -66,7 +66,7 @@ public class EthAccountsBodyProviderTest {
   }
 
   @Test
-  public void isParamIsAnObjectErrorIsReturned() {
+  public void ifParamIsAnObjectErrorIsReturned() {
     final String address = "MyAddress";
     final int id = 1;
     final EthAccountsBodyProvider bodyProvider = new EthAccountsBodyProvider(() -> Set.of(address));
